refactor: split `utils` module into meaningful modules
