+++ src/bridgewatcher/discord/formatting/__init__.py
@@ -1,0 +1,5 @@
+from .text import Markdown, format_number, readable_timestamp
+
+md = Markdown()
+
+__all__ = ("md", "format_number", "readable_timestamp")
