+++ src/bridgewatcher/discord/items/guesser.py
@@ -1,40 +1,11 @@
-from datetime import datetime
-from typing import Any
 from re import IGNORECASE, compile
 
 from discord import Interaction
 
-from bridgewatcher.api.model import Qualities
 from bridgewatcher.db import db
 from bridgewatcher.db.schema import Item, ItemName
 from bridgewatcher.discord.views import ItemPickerView
 from bridgewatcher.util.exc import NoItemFoundError, UntrackedItemRequested
-
-
-class Markdown:
-    def italic(self, s: Any) -> str:
-        return f"*{str(s)}*"
-
-    def bold(self, s: Any) -> str:
-        return f"**{str(s)}**"
-
-    def bold_italic(self, s: Any) -> str:
-        return f"***{str(s)}***"
-
-    def inline_code(self, s: Any) -> str:
-        return f"`{str(s)}`"
-
-    def underline(self, s: Any) -> str:
-        return f"__{str(s)}__"
-
-    def strikethrough(self, s: Any) -> str:
-        return f"~~{str(s)}~~"
-
-    def spoiler(self, s: Any) -> str:
-        return f"||{str(s)}||"
-
-    def link(self, display: str, url: str) -> str:
-        return f"[{display}]({url})"
 
 
 class ItemGuesser:
@@ -84,17 +55,3 @@ class ItemGuesser:
                 f"{item_name.name} is not stored in the database"
             )
         return Item.from_mongo(item), item_name
-
-
-def readable_timestamp(timestamp: str) -> str:
-    time = datetime.strptime(timestamp, "%Y-%m-%dT%H:%M:%S")
-    return f"{time.day:02d}/{time.month:02d}/{time.year} at {time.hour:02d}:{time.minute:02d}"
-
-
-def format_number(n: int) -> str:
-    return f"{n:,}"
-
-
-# not sure if this function belongs to this file
-def get_item_icon(item_id: str, quality: Qualities = Qualities.NORMAL) -> str:
-    return f"https://render.albiononline.com/v1/item/{item_id}?quality={quality.value}"
