--- src/bridgewatcher/discord/util/__init__.py
@@ -1,9 +1,0 @@
-from .server import ServerManager
-from .text import Markdown
-
-md = Markdown()
-
-__all__ = (
-    "md",
-    "ServerManager",
-)
